+++ lib/core/view_model/demo_view_model.py
@@ -1,0 +1,6 @@
+from lib.core.sdk.viewmodel import BaseViewModel
+
+
+class DemoViewModel(BaseViewModel):
+    test: str | None = None
+    carahlo: str | None = None
